GestoreMuseiPRO/requirements.txt
matplotlib==3.4.2
opencv_python==4.5.2.54
numpy==1.20.3
password_strength==0.0.3.post2
PyQRCode==1.2.1
pypng==0.0.20
PyQt5==5.15.4
qr==0.6.0
